requirements.txt
torch==2.6.0
datasets==3.1.0
transformers==4.51.3
vllm==0.8.5
openai==1.99.1
python-dotenv==1.1.1